+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.3 - 2020/10/13
+- Added typing (dropping py < 3.7)
+- Updated tables parser for pandoc > 2.10 (2020-06-29)
+- Removed cli2gui (this lib should be a dependency for that)
+
 ## 2020.2.1 - 2020/05/06
 - Updated classifiers